requirements.txt
asgiref==3.5.2
black==22.3.0
click==8.1.3
distlib==0.3.4
Django==4.0.4
django-rest-framework==0.1.0
djangorestframework==3.13.1
filelock==3.7.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
virtualenv==20.14.1
